Refatoração geral: unificação da utilização do `Result` como estrutura de resposta, ajustes nos retornos de handlers e controllers, inclusão de métodos de extensão em `ResultExtensions`, substituição de namespaces relacionados ao `UnitOfWork` e melhorias na consistência de erros com suporte a `ErrorCode`.

+++ src/02-Application/Application/Features/Catalog/Commands/Lesson/DeleteLesson/DeleteLessonCommandHandler.cs
@@ -1,5 +1,5 @@
-﻿using Application.Contracts.UnitOfWork;
-using Application.Interfaces.Repositories.Catalog;
+﻿using Application.Interfaces.Repositories.Catalog;
+using Application.Interfaces.UnitOfWork;
 using Domain.Commom;
 using MediatR;
 
@@ -13,6 +13,10 @@ public class DeleteLessonCommandHandler(ILessonRepository lessonRepository, IUni
 
     public async Task<Result> Handle(DeleteLessonCommand command, CancellationToken cancellationToken)
     {
+        var lesson = await _lessonRepository.GetById(command.Id);
+        if (lesson == null)
+            return Result.Failure("Lesson not found", "Not Found");
+        
         await _lessonRepository.Delete(command.Id);
         await _unitOfWork.CommitAsync();
         return Result.Success();
