+++ src/02-Application/Application/Features/Catalog/Commands/Module/DeleteModule/DeleteModuleCommandHandler.cs
@@ -1,5 +1,5 @@
-﻿using Application.Contracts.UnitOfWork;
-using Application.Interfaces.Repositories.Catalog;
+﻿using Application.Interfaces.Repositories.Catalog;
+using Application.Interfaces.UnitOfWork;
 using Domain.Commom;
 using MediatR;
 
@@ -10,6 +10,10 @@ public class DeleteModuleCommandHandler(IModuleRepository moduleRepository, IUni
 {
     public Task<Result> Handle(DeleteModuleCommand command, CancellationToken cancellationToken)
     {
+        var module = moduleRepository.GetById(command.Id);
+        if (module == null)
+            return Task.FromResult(Result.Failure("Module not found", "Not Found"));
+        
         moduleRepository.Delete(command.Id);
         unitOfWork.CommitAsync();
         return Task.FromResult(Result.Success());
